+++ main.go
@@ -44,7 +44,7 @@ import (
 )
 
 // Version 应用版本号
-var Version = "3.73.1"
+var Version = "3.73.2"
 
 // capitalDataSourceAdapter 资金數據源适配器
 type capitalDataSourceAdapter struct {